+++ src/protocol/solidircd.c
@@ -764,7 +764,7 @@ void solidircd_cmd_svsadmin(char *server, int set)
 /* SGLINE */
 void solidircd_cmd_sgline(char *mask, char *reason)
 {
-    send_cmd(NULL, "SGLINE %d :%s:%s", strlen(mask), mask, reason);
+    send_cmd(NULL, "SGLINE %d :%s:%s", (int)strlen(mask), mask, reason);
 }
 
 /* RAKILL */
